# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  controllers.CreateFaseBuahInput:
    properties:
      estimasi_panen:
        example: "2025-12-20"
        type: string
      jumlah_cover:
        example: 15
        type: integer
      minggu_ke:
        example: 3
        type: integer
      tanaman_id:
        example: 5
        type: integer
      tanggal_catat:
        example: "2025-11-29"
        type: string
      tanggal_cover:
        example: "2025-11-30"
        type: string
      warna_label:
        example: Hijau
        type: string
    type: object
...
  utils.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      meta: {}
      success:
        type: boolean
    type: object
host: localhost:2005
info:
  contact: {}
  description: Dokumentasi API Avocycle (CRUD, Pagination, dan Response Format)
  title: Avocycle API
  version: "1.0"
paths:
...
  /petani/fase-berbuah:
    get:
      description: Mendapatkan semua data fase berbuah dengan pagination
      parameters:
      - default: 1
        description: Halaman
        in: query
        name: page
        type: integer
      - default: 10
        description: Jumlah item per halaman
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Bearer: []
      summary: Get all fase berbuah
      tags:
      - Fase Berbuah
    post:
      consumes:
      - application/json
      description: Menambahkan data fase berbuah baru untuk tanaman
      parameters:
      - description: Fase Buah Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateFaseBuahInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Bearer: []
      summary: Create fase berbuah
      tags:
      - Fase Berbuah
  /petani/fase-berbuah/{id}:
    delete:
      description: Menghapus fase berbuah berdasarkan ID
      parameters:
      - description: ID Fase Buah
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Bearer: []
      summary: Delete fase berbuah
      tags:
      - Fase Berbuah
    get:
      description: Mendapatkan detail fase berbuah berdasarkan ID
      parameters:
      - description: ID Fase Buah
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Bearer: []
      summary: Get fase berbuah by ID
      tags:
      - Fase Berbuah
    put:
      consumes:
      - application/json
      description: Mengupdate data fase berbuah berdasarkan ID
      parameters:
      - description: ID Fase Buah
        in: path
        name: id
        required: true
        type: integer
      - description: Fase Buah Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateFaseBuahInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Bearer: []
      summary: Update fase berbuah
      tags:
      - Fase Berbuah
  /petani/fase-berbuah/tanaman/{tanaman_id}:
    get:
      description: Mendapatkan data fase berbuah berdasarkan tanaman ID dengan pagination
      parameters:
      - description: ID Tanaman
        in: path
        name: tanaman_id
        required: true
        type: integer
      - default: 1
        description: Halaman
        in: query
        name: page
        type: integer
      - default: 10
        description: Jumlah item per halaman
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Bearer: []
      summary: Get fase berbuah by tanaman ID
      tags:
      - Fase Berbuah
...
schemes:
- http
securityDefinitions:
  Bearer:
    description: Type "Bearer <your-token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"